Milestone 4: User Authentication

+++ resources/collection_dates.py
@@ -5,6 +5,7 @@ Blueprint for handling collection dates
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt
 
 from db import db
 from models import CollectionDateModel
@@ -23,6 +24,7 @@ class CollectionDates(MethodView):
     """
     Class for handling requests to the /collection_dates endpoint
     """
+    @jwt_required()
     @blp.response(200, CollectionDateSchema(many=True))
     def get(self):
         """
@@ -32,8 +34,18 @@ class CollectionDates(MethodView):
             dict: A dictionary containing all collection dates
             in the database
         """
-        return CollectionDateModel.query.all()
+        jwt = get_jwt()
 
+        user_role = jwt.get("role")
+
+        if user_role in ("admin", "household"):
+            return CollectionDateModel.query.all()
+
+        if user_role == "collector":
+            return CollectionDateModel.query.filter_by(
+                collector_id=jwt.get("sub")).all()
+
+    @jwt_required()
     @blp.arguments(CollectionDateSchema)
     @blp.response(201, CollectionDateSchema)
     def post(self, collection_date_data):
@@ -51,6 +63,14 @@ class CollectionDates(MethodView):
             abort(400, message): If there is an error adding the collection
             date to the database.
         """
+        jwt = get_jwt()
+
+        if jwt.get("role") != "collector":
+            abort(
+                403,
+                message="Collector privilege required to add collection dates"
+                )
+
         collection_date = CollectionDateModel(**collection_date_data)
 
         try:
@@ -69,20 +89,29 @@ class CollectionDate(MethodView):
     Class for handling requests to the /collection_dates/<collection_date_id>
     endpoint
     """
+    @jwt_required()
     @blp.response(200, CollectionDateSchema)
     def get(self, collection_date_id):
         """
         Get a collection date by ID
 
         Args:
-            collection_date_id (str): The ID of the collection date to retrieve
+            collection_date_id (str): ID of the collection date to retrieve
 
         Returns:
             tuple: A tuple containing the collection date and the HTTP
             status code 200
         """
-        return CollectionDateModel.query.get_or_404(collection_date_id)
+        jwt = get_jwt()
+        if jwt.get("role") in ("collector", "admin"):
+            return CollectionDateModel.query.get_or_404(collection_date_id)
 
+        abort(
+            403,
+            message="Collector privileges required to view collection dates"
+            )
+
+    @jwt_required()
     def delete(self, collection_date_id):
         """
         Delete a collection date by ID
@@ -99,6 +128,13 @@ class CollectionDate(MethodView):
             does not exist
 
         """
+        jwt = get_jwt()
+        if jwt.get("role") != "admin":
+            abort(
+                403,
+                message="Admin privileges required to delete collection dates"
+                )
+
         collection_date = CollectionDateModel.query.get_or_404(
             collection_date_id)
         db.session.delete(collection_date)
